cleanup and refactor (#16)
* refactor frontend

* refactor backend

* format

* add loading spinner & protect routes if not logged in

* add more unit tests

* format
