+++ backend/deno.json
@@ -11,8 +11,7 @@
 		"@tajpouria/cors": "jsr:@tajpouria/cors@^1.2.1",
 		"@std/assert": "jsr:@std/assert@1",
 		"@officeparser": "npm:officeparser",
-		"@djwt": "https://deno.land/x/djwt@v3.0.2/mod.ts",
-		"@dotenv": "https://deno.land/x/dotenv@v3.2.2/mod.ts"
+		"@djwt": "https://deno.land/x/djwt@v3.0.2/mod.ts"
 	},
 	"fmt": {
 		"useTabs": true
