+++ backend/util/generate_hash.test.ts
