+++ backend/util/generate_hash.ts
