+++ backend/util/util.test.ts
@@ -1,5 +1,6 @@
 import { assert } from "@std/assert";
 
+// create a body for use in tests
 export function createBody(body: string) {
 	const req = new Request("http://0.0.0.0", {
 		method: "POST",
