--- frontend/app/counter.module.css
@@ -1,4 +1,0 @@
-.buttons {
-	display: flex;
-	gap: 0.5rem;
-}
