--- frontend/app/counter.test.tsx
@@ -1,9 +1,0 @@
-import { fireEvent, render, screen } from "@testing-library/react";
-import Counter from "./counter";
-
-it("should increment when clicked", () => {
-	render(<Counter />);
-	expect(screen.getByRole("heading")).toHaveTextContent("0");
-	fireEvent.click(screen.getByRole("button", { name: "more" }));
-	expect(screen.getByRole("heading")).toHaveTextContent("1");
-});
