+++ frontend/app/register/register.tsx
@@ -9,7 +9,7 @@ import {
 	Input,
 } from "@mui/material";
 import { useRouter } from "next/navigation";
-import { useState } from "react";
+import { FormEvent, useState } from "react";
 import { backendPost } from "util/fetching";
 
 export function Register() {
@@ -18,99 +18,81 @@ export function Register() {
 	const [passMessage, setPassMessage] = useState("");
 	const [isLoading, setLoading] = useState(false);
 	const router = useRouter();
+	function onSubmit(event: FormEvent) {
+		event.preventDefault();
+		const data = new FormData(event.target as HTMLFormElement);
+
+		const email = data.get("email")?.toString() ?? "";
+		const username = data.get("username")?.toString() ?? "";
+		const password = data.get("password")?.toString() ?? "";
+		const confirmPassword = data.get("confirm_password")?.toString() ?? "";
+
+		// verify fields are not empty
+		if (!email || !username || !password || !confirmPassword) {
+			setBlankMessage(
+				"Please make sure you didn't leave any of the fields blank.",
+			);
+			return;
+		}
+		setBlankMessage("");
+
+		// verify passwords match
+		if (password != confirmPassword) {
+			setPassMessage("Please make sure passwords match.");
+			return;
+		}
+		setPassMessage("");
+
+		// input is valid
+		setLoading(true);
+		backendPost("api/register", {
+			email,
+			username,
+			password,
+		})
+			.then((data) => {
+				setPostData(data.message);
+				setLoading(false);
+				router.push("/login");
+			})
+			.catch((reason) => {
+				setPostData("Error: " + (reason?.message ?? reason));
+				setLoading(false);
+			});
+	}
 	return (
-		<>
-			<Card>
-				<CardHeader component="h2" title="Register" />
-				{isLoading ? <CircularProgress /> : undefined}
-				<CardContent>
-					<p data-testid="backend-register-post">{postData}</p>
-					<form
-						encType="text/plain"
-						onSubmit={(event) => {
-							event.preventDefault();
-							const data = new FormData(
-								event.target as HTMLFormElement,
-							);
-							// read fields
-							const fields = {
-								email: data.get("email")?.toString() ?? "",
-								username:
-									data.get("username")?.toString() ?? "",
-								password:
-									data.get("password")?.toString() ?? "",
-								confirmPassword:
-									data.get("confirm_password")?.toString() ??
-									"",
-							};
-
-							// verify fields are not empty
-							let isBlank = false;
-							for (const [_key, value] of Object.entries(
-								fields,
-							)) {
-								if (!value) isBlank = true;
-							}
-							if (isBlank)
-								setBlankMessage(
-									"Please make sure you didn't leave any of the fields blank.",
-								);
-							else setBlankMessage("");
-
-							// verify passwords match
-							const isMatching =
-								fields["password"] == fields["confirmPassword"];
-							if (isMatching) setPassMessage("");
-							else
-								setPassMessage(
-									"Please make sure passwords match.",
-								);
-
-							if (isMatching && !isBlank) {
-								setLoading(true);
-								// input is valid
-								backendPost("api/register", fields)
-									.then((data) => {
-										setLoading(false);
-										setPostData(data.message);
-										router.push("/login");
-									})
-									.catch((reason) => {
-										setLoading(false);
-										setPostData("" + reason);
-									});
-							}
-						}}
-					>
-						<Input name="email" placeholder="email" />
-						<br />
-						<Input name="username" placeholder="username" />
-						<br />
-						<Input
-							name="password"
-							type="password"
-							placeholder="password"
-						/>
-						<br />
-						<Input
-							name="confirm_password"
-							type="password"
-							placeholder="confirm password"
-						/>
-						{/* will let the user know when the passwords dont match*/}
-						<label data-testid="pass-message">{passMessage}</label>
-						<br />
-						{/* will let the user know when that fields cannot be left blank*/}
-						<label data-testid="blank-message">
-							{blankMessage}
-						</label>
-						<br />
-						<Button variant="contained" type="submit">
-							register
-						</Button>
-					</form>
-				</CardContent>
-			</Card>
-		</>
+		<Card>
+			<CardHeader component="h2" title="Register" />
+			{isLoading ? <CircularProgress /> : undefined}
+			<CardContent>
+				<p data-testid="backend-register-post">{postData}</p>
+				<form encType="text/plain" onSubmit={onSubmit}>
+					<Input name="email" placeholder="email" />
+					<br />
+					<Input name="username" placeholder="username" />
+					<br />
+					<Input
+						name="password"
+						type="password"
+						placeholder="password"
+					/>
+					<br />
+					<Input
+						name="confirm_password"
+						type="password"
+						placeholder="confirm password"
+					/>
+					{/* will let the user know when the passwords dont match*/}
+					<span data-testid="pass-message">{passMessage}</span>
+					<br />
+					{/* will let the user know when that fields cannot be left blank*/}
+					<span data-testid="blank-message">{blankMessage}</span>
+					<br />
+					<Button variant="contained" type="submit">
+						register
+					</Button>
+				</form>
+			</CardContent>
+		</Card>
 	);
 }
